# docs/swagger.yaml
definitions:
...
  github_com_ralph_industrial-edge-middleware_internal_models.Tag:
    properties:
      alias:
        type: string
      code:
        type: string
      created_at:
        type: string
      data_type:
        type: string
      gateway_id:
        type: integer
      historize:
        type: boolean
      historize_deadband:
        type: number
      id:
        type: integer
      sort_order:
        type: number
    type: object
...
  internal_handlers.CreateTagRequest:
    properties:
      alias:
        type: string
      code:
        type: string
      data_type:
        type: string
      gateway_id:
        type: integer
      historize:
        type: boolean
      historize_deadband:
        type: number
    required:
    - alias
    - code
    - data_type
    - gateway_id
    type: object
...
  internal_handlers.UpdateTagRequest:
    properties:
      alias:
        type: string
      code:
        type: string
      data_type:
        type: string
      historize:
        type: boolean
      historize_deadband:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /api/tags:
    get:
      consumes:
      - application/json
      description: Get a list of tags for the specified gateway
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Gateway ID
        in: query
        name: gateway_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ralph_industrial-edge-middleware_internal_models.Tag'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Gateway not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List tags
      tags:
      - tags
    post:
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Tag creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateTagRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ralph_industrial-edge-middleware_internal_models.Tag'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Gateway not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
  /api/tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tag by ID
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Tag deleted
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tag not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a tag
      tags:
      - tags
    get:
      consumes:
      - application/json
      description: Get a single tag by ID
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ralph_industrial-edge-middleware_internal_models.Tag'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tag not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a tag
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Update a tag by ID
      parameters:
      - description: Organization ID
        in: header
        name: X-Organization-ID
        required: true
        type: integer
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ralph_industrial-edge-middleware_internal_models.Tag'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Tag not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a tag
      tags:
      - tags
...
swagger: "2.0"